🧪 Add BaseFunctionalTest class for functional tests
This commit adds the `BaseFunctionalTest` class to the `CleanArchitecture.Api.FunctionalTests.Infrastructure` namespace. The `BaseFunctionalTest` class is an abstract class that serves as the base class for functional tests in the project. It provides a `httpClient` instance and a `GetToken` method for retrieving an authentication token. This class will be used as the base for other functional test classes.